+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Added Object.isPrimitiveObject(obj)
 * Added String.serializeAttributes(obj)
 * Added String#encodeHTML and String#decodeHTML
+* Added Object.flatten(obj)
 
 ## 0.1.1 (2014-06-23)